active_learning/requirements.txt
requests
pydantic
label-studio-sdk
typer
wandb
pyyaml
